move playground components into explicit playground folder
