+++ src/playground/components/allQuestions.tsx
@@ -1,9 +1,5 @@
-import React from 'react';
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import Grid from '@material-ui/core/Grid';
-import { Typography } from '@material-ui/core';
 import '@fontsource/montserrat';
+import { makeStyles, createStyles, Theme, Paper, Grid, Typography } from '@material-ui/core';
 
 const useStyles = makeStyles((theme: Theme) =>
 	createStyles({
@@ -24,7 +20,7 @@ const useStyles = makeStyles((theme: Theme) =>
 	})
 );
 
-export default function Question() {
+export default function AllQs() {
 	const classes = useStyles();
 
 	return (
